refactor: fold trace_to_sample into push.py, drop samples.py
`trace_to_sample` had no other in-repo consumer, so move it next to its
only caller (`push_traces`) and delete `verifiers/v1/samples.py`.
